+++ calliope/test/test_analysis.py
@@ -95,21 +95,3 @@ class TestModel:
         # TODO this should be tested with a more complex model
         unmet = analysis.get_unmet_demand_hours(builtin_model.solution)
         assert unmet == 1
-
-    def test_recompute_levelized_costs(self, model):
-        # Cost in solution
-        sol = model.solution
-        assert_almost_equal(sol['summary'].to_pandas().loc['ccgt', 'levelized_cost_monetary'], 0.1)
-        # Recomputed cost must be the same
-        dm = analysis.SolutionModel(model.solution)
-        recomputed = dm.recompute_levelized_costs('ccgt')
-        assert_almost_equal(recomputed['total'], 0.1)
-
-    def test_recompute_levelized_costs_after_changes(self, model):
-        # Make changes
-        dm = analysis.SolutionModel(model.solution)
-        dm.config_model.techs.ccgt.costs.monetary.e_cap = 50
-        dm.config_model.techs.ccgt.costs.monetary.om_fuel = 1.0
-        # Recomputed cost
-        recomputed = dm.recompute_levelized_costs('ccgt')
-        assert_almost_equal(recomputed['total'], 1.0, tolerance=0.001)